feat: event, listener, and notification for user approved event, sidebar condition for each  user types, overide authenticated in logincontroller

+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use App\Events\UserModerationApproved;
 use Illuminate\Support\Facades\Storage;
 use Yajra\DataTables\Facades\DataTables;
 use Illuminate\Support\Facades\Validator;
@@ -21,12 +22,13 @@ class UserController extends Controller
         return view('super.users.index');
     }
 
-    public function getUsers(Request $request) {
-        $users = User::latest()->get();
-            return DataTables::of($users)
-            ->addColumn('action', function($row) {
-                $actionBtn = '<a class="btn btn-primary btn-sm" href="'.route("users.edit",["user"=>$row->username]).'">Edit</a>
-                <button type="button" class="btn btn-danger text-white btn-sm delete-button" data-id="'.$row->username.'" id="btnDelete">
+    public function getUsers(Request $request)
+    {
+        $users = User::approved()->exceptlogged()->latest()->get();
+        return DataTables::of($users)
+            ->addColumn('action', function ($row) {
+                $actionBtn = '<a class="btn btn-primary btn-sm" href="' . route("users.edit", ["user" => $row->username]) . '">Edit</a>
+                <button type="button" class="btn btn-danger text-white btn-sm delete-button" data-id="' . $row->username . '" id="btnDelete">
                     Hapus
                 </button>';
                 return $actionBtn;
@@ -35,106 +37,128 @@ class UserController extends Controller
             ->make(true);
     }
 
-    public function indexApprove() {
+    public function indexApprove()
+    {
         return view('super.approve.index');
     }
 
-    public function getApproveUsers(Request $request) {
+    public function getApproveUsers(Request $request)
+    {
         $users = User::toapprove()->latest()->get();
-            return DataTables::of($users)
-            ->editColumn('verifikasi_file', function($row){
-                $raw = '<a href="'.Storage::url('file-verifikasi/'.$row->verifikasi_file).'" data-lightbox="'.$row->verifikasi_file.'" data-title="'.$row->username.'">'.$row->verifikasi_file.'</a>';
+        return DataTables::of($users)
+            ->editColumn('verifikasi_file', function ($row) {
+                $raw = '<a href="' . Storage::url('file-verifikasi/' . $row->verifikasi_file) . '" data-lightbox="' . $row->verifikasi_file . '" data-title="' . $row->username . '">' . $row->verifikasi_file . '</a>';
                 return $raw;
             })
-            ->addColumn('action', function($row) {
-                $actionBtn = '<button class="btn btn-success btn-sm approve-button" data-id="'.$row->username.'" id="btnApprove">Approve</button>
-                <button type="button" class="btn btn-danger text-white btn-sm reject-button" data-id="'.$row->username.'" id="btnReject">
+            ->addColumn('action', function ($row) {
+                $actionBtn = '<button class="btn btn-success btn-sm approve-button" data-id="' . $row->username . '" id="btnApprove">Approve</button>
+                <button type="button" class="btn btn-danger text-white btn-sm reject-button" data-id="' . $row->username . '" id="btnReject">
                     Reject
                 </button>';
                 return $actionBtn;
             })
-            ->rawColumns(['verifikasi_file','action'])
+            ->rawColumns(['verifikasi_file', 'action'])
             ->make(true);
     }
 
-    public function setApprovedUser(Request $request) {
+    public function setApprovedUser(Request $request)
+    {
         try {
             $user = User::findOrFail($request->username);
-            $destination = 'public/file-verifikasi/';
-            Storage::delete($destination . $user->verifikasi_file);
-            $user->verifikasi_file = '';
-            $user->terverifikasi = true;
-            $user->save();
-            return response()->json(['success' => 'Berhasil mengapprove user','data' => ['username' => $request->username]],200);
+            // $destination = 'public/file-verifikasi/';
+            // Storage::delete($destination . $user->verifikasi_file);
+            // $user->verifikasi_file = '';
+            // $user->terverifikasi = true;
+            // $user->save();
+            // event(new UserModerationApproved($user));
+            // Mail::to($user->email)->send(new UserModerationApprovedMail($user));
+            return response()->json(['success' => 'Berhasil mengapprove user', 'data' => ['id' => $user->hash_username]], 200);
         } catch (\Exception $e) {
-            return response()->json(['errors' => $e->getMessage()],500);
+            return response()->json(['errors' => $e->getMessage()], 500);
         }
     }
 
-    public function setRejectedUser(Request $request) {
+    public function sendEmailApproved(Request $request)
+    {
+        try {
+            $username = decryptString($request->id);
+            $user = User::findOrFail($username);
+            event(new UserModerationApproved($user));
+            return response()->json(['success' => 'Berhasil mengirimkan email']);
+        } catch (\Exception $e) {
+            return response()->json(['errors' => $e->getMessage()], 500);
+        }
+    }
+
+    public function setRejectedUser(Request $request)
+    {
         try {
             $user = User::findOrFail($request->username);
             $destination = 'public/file-verifikasi/';
             Storage::delete($destination . $user->verifikasi_file);
             $user->delete();
-            return response()->json(['success' => 'Berhasil menolak dan menghapus data user','data' => ['username' => $request->username]],200);
+            return response()->json(['success' => 'Berhasil menolak dan menghapus data user', 'data' => ['username' => $request->username]], 200);
         } catch (\Exception $e) {
-            return response()->json(['errors' => $e->getMessage()],500);
+            return response()->json(['errors' => $e->getMessage()], 500);
         }
     }
 
-    public function profile() {
+    public function profile()
+    {
         $user = auth()->user();
         return view('akun.profile.index', compact('user'));
     }
-    
-    public function updateProfile(Request $request) {
+
+    public function updateProfile(Request $request)
+    {
         $user = Auth::user();
         $validator = Validator::make($request->all(), [
             'nama' => 'required|string|min:2|max:150',
             'email' => [
-                'required','email',
+                'required', 'email',
                 Rule::unique('users')->ignore($user->username, 'username')
             ]
         ]);
 
-        if($validator->fails()) {
-            return response()->json(['errors'=> $validator->errors()],422);
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors()], 422);
         }
         $data = $validator->validated();
         $user->nama = $data['nama'];
         $user->email = $data['email'];
         $user->save();
-        
-        return response()->json(['success' => 'Berhasil mengupdate profile','data'=> ['nama' => $data['nama']]]);
+
+        return response()->json(['success' => 'Berhasil mengupdate profile', 'data' => ['nama' => $data['nama']]]);
     }
 
-    public function security() {
+    public function security()
+    {
         return view('akun.keamanan.index');
     }
-    
-    public function securityUpdate(Request $request) {
+
+    public function securityUpdate(Request $request)
+    {
         $validator = Validator::make($request->all(), [
             'current_password' => 'required',
             'password_confirmation' => 'required',
             'password' => 'required|confirmed|min:8|different:current_password',
         ]);
 
-        if($validator->fails()){
-            return response()->json(['errors' => $validator->errors()],422);
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors()], 422);
         }
         $data = $validator->validated();
         $user = Auth::user();
-        if(Hash::check($data['current_password'], $user->password)){
+        if (Hash::check($data['current_password'], $user->password)) {
             $user->fill([
                 'password' => Hash::make($data['password'])
             ])->save();
-            
+
             return response()->json(['success' => 'Berhasil mengupdate password']);
         }
         return response()->json(['errors' => ['current_password' => [
             "Password does not match"
-        ]]],422);
+        ]]], 422);
     }
 
     /**
@@ -158,8 +182,8 @@ class UserController extends Controller
             'role' => 'required|in:super,admin,user'
         ]);
 
-        if($validator->fails()) {
-            return response()->json(['errors'=> $validator->errors()],422);
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors()], 422);
         }
         $validData = $validator->validated();
         $user = User::create([
@@ -171,7 +195,7 @@ class UserController extends Controller
             'role' => $validData['role'],
         ]);
 
-        return response()->json(['success' => 'Berhasil menambahkan pengguna','data'=>$user],200);
+        return response()->json(['success' => 'Berhasil menambahkan pengguna', 'data' => $user], 200);
     }
 
     /**
@@ -187,6 +211,10 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $loggedInUserId = auth()->user()->username;
+        if ($user->username === $loggedInUserId) {
+            return abort(404);
+        }
         return view('super.users.form-user', compact('user'));
     }
 
@@ -198,7 +226,7 @@ class UserController extends Controller
         $validator = Validator::make($request->all(), [
             'nama' => 'required|string|min:2|max:150',
             'email' => [
-                'required','email',
+                'required', 'email',
                 Rule::unique('users')->ignore($user->username, 'username')
             ],
             'username' => [
@@ -212,7 +240,7 @@ class UserController extends Controller
             'role' => 'required|in:super,admin,user'
         ]);
 
-        if($validator->fails()) {
+        if ($validator->fails()) {
             return redirect()->back()->withErrors($validator)->withInput();
         }
 
@@ -236,26 +264,33 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-        try{
+        try {
+            $loggedInUserId = auth()->user()->username;
+            if ($user->username === $loggedInUserId) {
+                throw new \Exception('User Not Found!', 404);
+            }
             $user->delete();
-            return response()->json(['success' => 'Berhasil menghapus data'],200);
-        }catch(\Exception $e) {
-            return response()->json(['errors' => $e->getMessage()],422);
+            return response()->json(['success' => 'Berhasil menghapus data'], 200);
+        } catch (\Exception $e) {
+            return response()->json(['errors' => $e->getMessage()], 422);
         }
     }
 
-    public function profileLanding(){
+    public function profileLanding()
+    {
         $user = Auth::user();
-        return view ('profile',compact('user'));
+        return view('profile', compact('user'));
     }
 
-    public function settingLanding(){
+    public function settingLanding()
+    {
         $user = Auth::user();
-        return view ('setting.profile',compact('user'));
+        return view('setting.profile', compact('user'));
     }
 
-    public function keamananLanding(){
+    public function keamananLanding()
+    {
         $user = Auth::user();
-        return view ('setting.keamanan',compact('user'));
+        return view('setting.keamanan', compact('user'));
     }
 }
